Build the 64-bit Linux port as a PIE and fix pointer truncation
The 64-bit build linked with -no-pie, so everything was loaded below
4 GB and pointers cut down to 32 bits happened to keep working. Build it
position independent instead, so any truncation fails right away.

- Link with -pie -z text, and keep assembled data (which holds absolute
  pointers) in .data.rel.ro so it needs no text relocations.
- Make int/pointer size mismatches errors for all PC builds.
- Fix the truncations that exposed: RLUnCompVram (contest paintings),
  the script address offset used by vgoto etc., the DMA address
  registers, hidden item data read through the script pointer, and a
  few casts.
- Define SPECIAL_* ids locally in every script file, since a 16-bit
  relocation against them isn't allowed in a PIE.

+++ src/battle_anim_fight.c
@@ -838,7 +838,7 @@ static void AnimSuperpowerRock(struct Sprite *sprite)
     sprite->y = 120;
 
     sprite->data[0] = gBattleAnimArgs[3];
-    StorePointerInVars(&sprite->ptr.intPtr, (void *)(sprite->y << 8));
+    StorePointerInVars(&sprite->ptr.intPtr, (void *)(intptr_t)(sprite->y << 8));
 
     sprite->data[6] = gBattleAnimArgs[1];
     sprite->oam.tileNum += gBattleAnimArgs[2] * 4;
